refactor: reorganize project structure
- Extract pure settings reads from AuthManager into config.ts (getBaseUrl,
  getApiModelId, getMaxTokens)
- Centralize all compile-time constants into consts.ts (9 constants from 5
  files), group by chapter with JSDoc
- Inline findModel helper
- Replace import * as vscode with default import (esModuleInterop)
- Remove MODELS from types.ts → consts.ts for cleaner type-only module
