+++ src/config.ts
@@ -1,0 +1,35 @@
+import vscode from 'vscode';
+import { CONFIG_SECTION } from './consts';
+
+/**
+ * Get DeepSeek API base URL from settings.
+ * Falls back to the official endpoint when not configured.
+ */
+export function getBaseUrl(): string {
+	const config = vscode.workspace.getConfiguration(CONFIG_SECTION);
+	return config.get<string>('baseUrl') || 'https://api.deepseek.com';
+}
+
+/**
+ * Resolve the API model ID to send to the endpoint.
+ *
+ * Users can override model IDs via the `modelIdOverrides` setting object
+ * (e.g. for third-party API proxies). Falls back to the VS Code model ID
+ * when no override is configured.
+ */
+export function getApiModelId(vscodeModelId: string): string {
+	const config = vscode.workspace.getConfiguration(CONFIG_SECTION);
+	const overrides = config.get<Record<string, string>>('modelIdOverrides');
+	const override = overrides?.[vscodeModelId]?.trim();
+	return override || vscodeModelId;
+}
+
+/**
+ * Get the configured max output tokens limit.
+ * Returns `undefined` when set to 0 (API default — no limit).
+ */
+export function getMaxTokens(): number | undefined {
+	const config = vscode.workspace.getConfiguration(CONFIG_SECTION);
+	const value = config.get<number>('maxTokens', 0);
+	return value > 0 ? value : undefined;
+}
