+++ src/types.ts
@@ -106,36 +106,3 @@ export interface ModelDefinition {
 	};
 	requiresThinkingParam: boolean;
 }
-
-export const MODELS: ModelDefinition[] = [
-	{
-		id: 'deepseek-v4-flash',
-		name: 'DeepSeek V4 Flash',
-		family: 'deepseek',
-		version: 'v4',
-		detail: 'Fast, general-purpose model',
-		maxInputTokens: 1048576,
-		maxOutputTokens: 393216,
-		capabilities: {
-			toolCalling: true,
-			imageInput: true,
-			thinking: true,
-		},
-		requiresThinkingParam: true,
-	},
-	{
-		id: 'deepseek-v4-pro',
-		name: 'DeepSeek V4 Pro',
-		family: 'deepseek',
-		version: 'v4',
-		detail: 'Most capable reasoning model',
-		maxInputTokens: 1048576,
-		maxOutputTokens: 393216,
-		capabilities: {
-			toolCalling: true,
-			imageInput: true,
-			thinking: true,
-		},
-		requiresThinkingParam: true,
-	},
-];
